Refactor runners into a separate module
