+++ docs/EmpireStateRunUp/empirestaterunup/apps.py
@@ -3,7 +3,6 @@ Collection of applications to display race findings
 author: Jose Vicente Nunez <kodegeek.com@protonmail.com>
 """
 import textwrap
-from argparse import ArgumentParser
 from pathlib import Path
 from typing import Type
 
@@ -14,7 +13,6 @@ from textual.containers import HorizontalScroll, VerticalScroll
 from textual.driver import Driver
 from textual.screen import ModalScreen
 from textual.widgets import DataTable, Footer, Header, Log, Label, Button, MarkdownViewer
-import matplotlib.pyplot as plt
 
 from empirestaterunup.analyze import SUMMARY_METRICS, get_5_number, count_by_age, count_by_gender, count_by_wave, \
     dt_to_sorted_dict, get_outliers, age_bins, time_bins
@@ -109,26 +107,6 @@ class FiveNumberApp(App):
         table = event.data_table
         if table.id != 'Summary':  # Not supported yet
             table.sort(event.column_key)
-
-
-def run_5_number():
-    parser = ArgumentParser(description="5 key indicators report")
-    parser.add_argument(
-        "results",
-        action="store",
-        type=Path,
-        nargs="*",
-        help="Race results."
-    )
-    options = parser.parse_args()
-    app = FiveNumberApp()
-    if options.results:
-        FiveNumberApp.DF = load_data(options.results[0])
-    else:
-        FiveNumberApp.DF = load_data()
-    app.title = f"Five Number Summary".title()
-    app.sub_title = f"Runners: {FiveNumberApp.DF.shape[0]}"
-    app.run()
 
 
 class RunnerDetailScreen(ModalScreen):
@@ -228,26 +206,6 @@ class OutlierApp(App):
         self.push_screen(runner_detail)
 
 
-def run_outlier():
-    parser = ArgumentParser(description="Show race outliers")
-    parser.add_argument(
-        "results",
-        action="store",
-        type=Path,
-        nargs="*",
-        help="Race results."
-    )
-    options = parser.parse_args()
-    if options.results:
-        OutlierApp.DF = load_data(options.results[0])
-    else:
-        OutlierApp.DF = load_data()
-    app = OutlierApp()
-    app.title = f"Outliers Summary".title()
-    app.sub_title = f"Runners: {OutlierApp.DF.shape[0]}"
-    app.run()
-
-
 class Plotter:
 
     def __init__(self, data_file: Path = None):
@@ -284,43 +242,6 @@ class Plotter:
             subplots=True,
             autopct="%.2f"
         )
-
-
-def simple_plot():
-    parser = ArgumentParser(description="Different Age plots for Empire State RunUp")
-    parser.add_argument(
-        "--type",
-        action="store",
-        default="box",
-        choices=["box", "hist"],
-        help="Plot type. Not all reports honor this choice (like country)"
-    )
-    parser.add_argument(
-        "--report",
-        action="store",
-        default="age",
-        choices=["age", "country", "gender"],
-        help="Report type"
-    )
-    parser.add_argument(
-        "results",
-        action="store",
-        type=Path,
-        nargs="*",
-        help="Race results."
-    )
-    options = parser.parse_args()
-    if options.results:
-        pzs = Plotter(options.results[0])
-    else:
-        pzs = Plotter()
-    if options.report == 'age':
-        pzs.plot_age(options.type)
-    elif options.report == 'country':
-        pzs.plot_country()
-    elif options.report == 'gender':
-        pzs.plot_gender()
-    plt.show()
 
 
 class BrowserApp(App):
@@ -383,33 +304,3 @@ class BrowserApp(App):
         table.sort(event.column_key)
 
 
-def run_browser():
-    parser = ArgumentParser(description="Browse user results")
-    parser.add_argument(
-        "--country",
-        action="store",
-        type=Path,
-        required=False,
-        help="Country details"
-    )
-    parser.add_argument(
-        "results",
-        action="store",
-        type=Path,
-        nargs="*",
-        help="Race results."
-    )
-    options = parser.parse_args()
-    df = None
-    country_df = None
-    if options.results:
-        df = load_data(options.results[0])
-    if options.country:
-        country_df = load_country_details(options.country)
-    app = BrowserApp(
-        df=df,
-        country_data=country_df
-    )
-    app.title = f"Race runners".title()
-    app.sub_title = f"Browse details: {app.df.shape[0]}"
-    app.run()
